rename State to Dataset; remove single and ensemble ccts
