+++ app/src/components/callflow.js
@@ -68,7 +68,7 @@ export default {
 		},
 		left: false,
 		formats: ["CCT", "SuperGraph"],
-		selectedFormat: "SuperGraph",
+		selectedFormat: "CCT",
 		datasets: [],
 		selectedTargetDataset: "",
 		selectedDataset2: "",
@@ -201,6 +201,7 @@ export default {
 			this.setTargetDataset();
 			this.setComponentMap();
 
+			console.log(this.selectedFormat.length, this.selectedMode)
 			if (this.selectedFormat == "SuperGraph") {
 				if (this.selectedMode == "Single") {
 					this.$socket.emit("single_callsite_data", {
@@ -266,13 +267,6 @@ export default {
 
 	methods: {
 		// Feature: Sortby the datasets and show the time. 
-		formatRuntimeWithoutUnits(val) {
-			let format = d3.format(".2");
-			let ret = format(val);
-			return ret;
-		},
-
-		// Feature: Sortby the datasets and show the time. 
 		sortDatasetsByAttr(datasets, attr) {
 			let ret = datasets.sort((a, b) => {
 				let x = 0, y = 0;
@@ -330,10 +324,10 @@ export default {
 				this.selectedMode = "Single";
 			}
 
-			this.$store.maxExcTime = data["max_excTime"];
-			this.$store.minExcTime = data["min_excTime"];
-			this.$store.maxIncTime = data["max_incTime"];
-			this.$store.minIncTime = data["min_incTime"];
+			this.$store.maxExcTime = data["maxExcTime"];
+			this.$store.minExcTime = data["minExcTime"];
+			this.$store.maxIncTime = data["maxIncTime"];
+			this.$store.minIncTime = data["minIncTime"];
 
 			this.$store.numOfRanks = data["numOfRanks"];
 			this.$store.moduleCallsiteMap = data["module_callsite_map"];
@@ -341,8 +335,6 @@ export default {
 
 			this.$store.selectedMPIBinCount = this.selectedMPIBinCount;
 			this.$store.selectedRunBinCount = this.selectedRunBinCount;
-
-			this.selectedIncTime = ((this.selectedFilterPerc * this.$store.maxIncTime[this.selectedTargetDataset] * 0.000001) / 100).toFixed(3);
 
 			this.setViewDimensions();
 
@@ -383,6 +375,7 @@ export default {
 				this.$store.resetTargetDataset = true;
 			}
 			this.$store.selectedMetric = this.selectedMetric;
+			console.log(this.$store.selectedDatasets)
 			this.datasets = this.sortDatasetsByAttr(this.$store.selectedDatasets, "Inclusive");
 
 			let max_dataset = "";
@@ -411,8 +404,9 @@ export default {
 			else {
 				this.$store.selectedTargetDataset = this.selectedTargetDataset;
 			}
+			this.selectedIncTime = ((this.selectedFilterPerc * this.$store.maxIncTime[this.selectedTargetDataset] * 0.000001) / 100).toFixed(3);
 
-			console.log("Minimum among all runtimes: ", this.selectedTargetDataset);
+			console.log("Maximum among all runtimes: ", this.selectedTargetDataset);
 		},
 
 		setComponentMap() {
@@ -595,7 +589,9 @@ export default {
 			this.setupColors();
 			this.setOtherData();
 			this.setTargetDataset();
-			this.setSelectedModule();
+			if(this.selectedFormat == 'SuperGraph'){
+				this.setSelectedModule();
+			}
 
 			console.log("Mode : ", this.selectedMode);
 			console.log("Number of runs :", this.$store.numOfRuns);
