+++ callflow/__init__.py
@@ -4,12 +4,16 @@ from .utils import *
 from .timer import Timer
 
 from .datastructures.graphframe import GraphFrame
+from .datastructures.dataset import Dataset
 
 from .datastructures.supergraph_ensemble import EnsembleSuperGraph
 from .datastructures.supergraph_single import SingleSuperGraph
-from .datastructures.cct_ensemble import EnsembleCCT
-from .datastructures.cct_single import SingleCCT
+from .datastructures.cct import CCT
+
 
 from .callflow_base import BaseCallFlow
 from .callflow_single import SingleCallFlow
 from .callflow_ensemble import EnsembleCallFlow
+
+
+
