+++ callflow/datastructures/graphframe.py
@@ -70,30 +70,10 @@ class GraphFrame(ht.GraphFrame):
     def lookup_with_vis_nodeName(self, name):
         return self.df.loc[self.df["vis_node_name"] == name]
 
-    """
-    def lookup_by_column(self, _hash, col_name):
-        ret = []
-        node_df = self.df.loc[self.df["node"] == self.map[str(_hash)]]
-        node_df_T = node_df.T.squeeze()
-        node_df_T_attr = node_df_T.loc[col_name]
-        if node_df_T_attr is not None:
-            if type(node_df_T_attr) is str or type(node_df_T_attr) is float:
-                ret.append(node_df_T_attr)
-            else:
-                ret = node_df_T_attr.tolist()
-        return ret
-    """
     # --------------------------------------------------------------------------
     def update_df(self, col_name, mapping):
         self.df[col_name] = self.df["name"].apply(
             lambda node: mapping[node] if node in mapping.keys() else ""
         )
-
-    def grouped_df(self, attr):
-        pass
-        """
-        self.gdf[attr] = self.df.groupby(attr, as_index=True, squeeze=True)
-        self.gdfKeys = self.gdf[attr].groups.keys()
-        """
 
     # --------------------------------------------------------------------------
