+++ callflow/pipeline/__init__.py
@@ -3,5 +3,4 @@ from .filter_networkx import FilterNetworkX
 from .group_by_module import groupBy
 from .process import PreProcess
 from .index import Pipeline
-from .state import State
 from .convert_hatchet_to_networkx import HatchetToNetworkX
